+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/implementation/cosmos/AzureCosmosAutoConfiguration.java
@@ -3,71 +3,19 @@
 
 package com.azure.spring.cloud.autoconfigure.implementation.cosmos;
 
-import com.azure.cosmos.CosmosAsyncClient;
-import com.azure.cosmos.CosmosClient;
-import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.spring.cloud.autoconfigure.implementation.AzureServiceConfigurationBase;
-import com.azure.spring.cloud.autoconfigure.implementation.condition.ConditionalOnAnyProperty;
-import com.azure.spring.cloud.autoconfigure.implementation.context.properties.AzureGlobalProperties;
-import com.azure.spring.cloud.autoconfigure.implementation.cosmos.properties.AzureCosmosProperties;
-import com.azure.spring.cloud.core.customizer.AzureServiceClientBuilderCustomizer;
-import com.azure.spring.cloud.core.implementation.util.AzureSpringIdentifier;
-import com.azure.spring.cloud.service.implementation.cosmos.CosmosClientBuilderFactory;
-import org.springframework.beans.factory.ObjectProvider;
+import com.azure.spring.cloud.autoconfigure.implementation.cosmos.properties.AzureCosmosPropertiesConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 
 /**
  * {@link EnableAutoConfiguration Auto-configuration} for Azure Cosmos DB support.
  *
  * @since 4.0.0
  */
-@ConditionalOnClass(CosmosClientBuilder.class)
-@ConditionalOnProperty(value = "spring.cloud.azure.cosmos.enabled", havingValue = "true", matchIfMissing = true)
-@ConditionalOnAnyProperty(prefix = "spring.cloud.azure.cosmos", name = "endpoint")
-public class AzureCosmosAutoConfiguration extends AzureServiceConfigurationBase {
-
-    AzureCosmosAutoConfiguration(AzureGlobalProperties azureGlobalProperties) {
-        super(azureGlobalProperties);
-    }
-
-    @Bean
-    @ConfigurationProperties(AzureCosmosProperties.PREFIX)
-    AzureCosmosProperties azureCosmosProperties() {
-        return loadProperties(getAzureGlobalProperties(), new AzureCosmosProperties());
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    CosmosClient azureCosmosClient(CosmosClientBuilder builder) {
-        return builder.buildClient();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    // TODO (xiada): spring data cosmos also defines a CosmosAsyncClient
-    CosmosAsyncClient azureCosmosAsyncClient(CosmosClientBuilder builder) {
-        return builder.buildAsyncClient();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    CosmosClientBuilder cosmosClientBuilder(CosmosClientBuilderFactory factory) {
-        return factory.build();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    CosmosClientBuilderFactory cosmosClientBuilderFactory(AzureCosmosProperties properties,
-        ObjectProvider<AzureServiceClientBuilderCustomizer<CosmosClientBuilder>> customizers) {
-        CosmosClientBuilderFactory factory = new CosmosClientBuilderFactory(properties);
-        factory.setSpringIdentifier(AzureSpringIdentifier.AZURE_SPRING_COSMOS);
-        customizers.orderedStream().forEach(factory::addBuilderCustomizer);
-        return factory;
-    }
+@Import({
+    AzureCosmosPropertiesConfiguration.class,
+    CosmosClientConfiguration.class
+})
+public class AzureCosmosAutoConfiguration {
 
 }