+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/implementation/storage/blob/AzureStorageBlobProtocolResolverConfiguration.java
@@ -1,0 +1,29 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.spring.cloud.autoconfigure.implementation.storage.blob;
+
+import com.azure.spring.cloud.autoconfigure.implementation.storage.blob.properties.AzureStorageBlobProperties;
+import com.azure.spring.cloud.core.resource.AzureStorageBlobProtocolResolver;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+
+/**
+ * {@link EnableAutoConfiguration Auto-configuration} for Spring Resource with Azure Storage Blob support.
+ *
+ *  @since 4.0.0
+ */
+@ConditionalOnClass({ AzureStorageBlobProtocolResolver.class })
+@ConditionalOnBean(AzureStorageBlobProperties.class)
+public class AzureStorageBlobProtocolResolverConfiguration {
+
+    @Bean
+    @ConditionalOnMissingBean
+    AzureStorageBlobProtocolResolver azureStorageBlobProtocolResolver() {
+        return new AzureStorageBlobProtocolResolver();
+    }
+
+}