+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/implementation/data/cosmos/CosmosDataAutoConfigurationTests.java
@@ -3,8 +3,10 @@
 
 package com.azure.spring.cloud.autoconfigure.implementation.data.cosmos;
 
+import com.azure.cosmos.ConnectionMode;
 import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.spring.cloud.autoconfigure.implementation.cosmos.properties.AzureCosmosProperties;
+import com.azure.spring.cloud.autoconfigure.implementation.context.properties.AzureGlobalProperties;
+import com.azure.spring.cloud.autoconfigure.implementation.cosmos.properties.AzureCosmosConnectionDetails;
 import com.azure.spring.data.cosmos.CosmosFactory;
 import com.azure.spring.data.cosmos.config.CosmosConfig;
 import com.azure.spring.data.cosmos.core.CosmosTemplate;
@@ -17,7 +19,9 @@ import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 
+import static com.azure.spring.cloud.autoconfigure.implementation.data.cosmos.CosmosDataAutoConfigurationTests.CustomAzureCosmosConnectionDetails.DATABASE;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.RETURNS_MOCKS;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
@@ -51,15 +55,51 @@ class CosmosDataAutoConfigurationTests {
 
     @Test
     @SuppressWarnings("try")
+    void onlyConnectionDetailsBeanConfigured() {
+        try (MockedStatic<CosmosFactory> ignored = mockStatic(CosmosFactory.class, RETURNS_MOCKS)) {
+            this.contextRunner
+                .withBean(AzureGlobalProperties.class, this::globalProperties)
+                .withBean(CosmosClientBuilder.class, () -> mock(CosmosClientBuilder.class))
+                .withPropertyValues(
+                    "spring.cloud.azure.cosmos.endpoint=" + ENDPOINT,
+                    "spring.cloud.azure.cosmos.database=property-database"
+                )
+                .withBean(AzureCosmosConnectionDetails.class, CustomAzureCosmosConnectionDetails::new)
+                .run(context -> {
+                    assertThat(context).hasSingleBean(CosmosTemplate.class);
+                    assertThat(context).hasSingleBean(ReactiveCosmosTemplate.class);
+                });
+        }
+    }
+
+    @Test
+    @SuppressWarnings("try")
+    void bothPropertyAndBeanConfiguredBeanHasHigherPriority() {
+        try (MockedStatic<CosmosFactory> ignored = mockStatic(CosmosFactory.class, RETURNS_MOCKS)) {
+            this.contextRunner
+                .withBean(AzureGlobalProperties.class, this::globalProperties)
+                .withBean(CosmosClientBuilder.class, () -> mock(CosmosClientBuilder.class))
+                .withPropertyValues(
+                    "spring.cloud.azure.cosmos.endpoint=" + ENDPOINT,
+                    "spring.cloud.azure.cosmos.database=property-database"
+                )
+                .withBean(AzureCosmosConnectionDetails.class, CustomAzureCosmosConnectionDetails::new)
+                .run(context -> {
+                    assertThat(context).hasSingleBean(CosmosTemplate.class);
+                    assertThat(context).hasSingleBean(ReactiveCosmosTemplate.class);
+                    assertThat(context).hasSingleBean(CosmosDataAutoConfiguration.class);
+                    assertEquals(DATABASE, context.getBean(CosmosDataConfiguration.class).getDatabaseName());
+                });
+        }
+    }
+
+    @Test
+    @SuppressWarnings("try")
     void cosmosTemplateExistsAndUsesIt() {
         try (MockedStatic<CosmosFactory> ignored = mockStatic(CosmosFactory.class, RETURNS_MOCKS)) {
-            AzureCosmosProperties azureCosmosProperties = new AzureCosmosProperties();
-            azureCosmosProperties.setEndpoint(ENDPOINT);
-            azureCosmosProperties.setDatabase("test");
-
             this.contextRunner
                 .withBean(CosmosClientBuilder.class, () -> mock(CosmosClientBuilder.class))
-                .withBean(AzureCosmosProperties.class, () -> azureCosmosProperties)
+                .withBean(AzureGlobalProperties.class, this::globalProperties)
                 .withPropertyValues(
                     "spring.cloud.azure.cosmos.endpoint=" + ENDPOINT,
                     "spring.cloud.azure.cosmos.database=test"
@@ -75,13 +115,9 @@ class CosmosDataAutoConfigurationTests {
     @SuppressWarnings("try")
     void noResponseDiagnosticsProcessorByDefault() {
         try (MockedStatic<CosmosFactory> ignored = mockStatic(CosmosFactory.class, RETURNS_MOCKS)) {
-            AzureCosmosProperties azureCosmosProperties = new AzureCosmosProperties();
-            azureCosmosProperties.setEndpoint(ENDPOINT);
-            azureCosmosProperties.setDatabase("test");
-
             this.contextRunner
                 .withBean(CosmosClientBuilder.class, () -> mock(CosmosClientBuilder.class))
-                .withBean(AzureCosmosProperties.class, () -> azureCosmosProperties)
+                .withBean(AzureGlobalProperties.class, this::globalProperties)
                 .withPropertyValues(
                     "spring.cloud.azure.cosmos.endpoint=" + ENDPOINT,
                     "spring.cloud.azure.cosmos.database=test"
@@ -98,14 +134,9 @@ class CosmosDataAutoConfigurationTests {
     @SuppressWarnings("try")
     void hasResponseDiagnosticsProcessorWhenEnableQueryMetrics() {
         try (MockedStatic<CosmosFactory> ignored = mockStatic(CosmosFactory.class, RETURNS_MOCKS)) {
-            AzureCosmosProperties azureCosmosProperties = new AzureCosmosProperties();
-            azureCosmosProperties.setEndpoint(ENDPOINT);
-            azureCosmosProperties.setDatabase("test");
-            azureCosmosProperties.setPopulateQueryMetrics(true);
-
             this.contextRunner
                 .withBean(CosmosClientBuilder.class, () -> mock(CosmosClientBuilder.class))
-                .withBean(AzureCosmosProperties.class, () -> azureCosmosProperties)
+                .withBean(AzureGlobalProperties.class, this::globalProperties)
                 .withPropertyValues(
                     "spring.cloud.azure.cosmos.endpoint=" + ENDPOINT,
                     "spring.cloud.azure.cosmos.database=test",
@@ -127,13 +158,9 @@ class CosmosDataAutoConfigurationTests {
     @Test
     void testWithEnableAuditing() {
         try (MockedStatic<CosmosFactory> ignored = mockStatic(CosmosFactory.class, RETURNS_MOCKS)) {
-            AzureCosmosProperties azureCosmosProperties = new AzureCosmosProperties();
-            azureCosmosProperties.setEndpoint(ENDPOINT);
-            azureCosmosProperties.setDatabase("test");
-
             this.contextRunner
                 .withBean(CosmosClientBuilder.class, () -> mock(CosmosClientBuilder.class))
-                .withBean(AzureCosmosProperties.class, () -> azureCosmosProperties)
+                .withBean(AzureGlobalProperties.class, this::globalProperties)
                 .withUserConfiguration(UserAuditingConfiguration.class)
                 .withPropertyValues(
                     "spring.cloud.azure.cosmos.endpoint=" + ENDPOINT,
@@ -149,5 +176,43 @@ class CosmosDataAutoConfigurationTests {
     @EnableCosmosAuditing
     static class UserAuditingConfiguration {
 
+    }
+
+    private AzureGlobalProperties globalProperties() {
+        AzureGlobalProperties globalProperties = new AzureGlobalProperties();
+        globalProperties.getCredential().setClientId("azure-client-id");
+        globalProperties.getCredential().setClientSecret("azure-client-secret");
+        globalProperties.getProxy().setHostname("localhost");
+        globalProperties.getProxy().getHttp().setNonProxyHosts("localhost");
+        return globalProperties;
+    }
+
+    static class CustomAzureCosmosConnectionDetails implements AzureCosmosConnectionDetails {
+        static final String DATABASE = "bean-database";
+
+        @Override
+        public String getEndpoint() {
+            return "bean-endpoint";
+        }
+
+        @Override
+        public String getKey() {
+            return "bean-key";
+        }
+
+        @Override
+        public String getDatabase() {
+            return DATABASE;
+        }
+
+        @Override
+        public Boolean getEndpointDiscoveryEnabled() {
+            return false;
+        }
+
+        @Override
+        public ConnectionMode getConnectionMode() {
+            return ConnectionMode.GATEWAY;
+        }
     }
 }